requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.1
Pillow==10.1.0
pdf2image==1.16.3
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
transformers
torch
pytest
